# Optimized aggregating losses