--- a/openquake/qa_tests_data/scenario_risk/case_1/expected/losses_by_event.csv
+++ b/openquake/qa_tests_data/scenario_risk/case_1/expected/losses_by_event.csv
@@ -1,4 +1,4 @@
-#,,"generated_by='OpenQuake engine 3.12.0-git8ca653b59f', start_date='2021-03-15T07:11:37', checksum=3590982611"
+#,,"generated_by='OpenQuake engine 3.12.0-git3180a4f5a5', start_date='2021-03-19T09:29:03', checksum=3590982611"
 event_id,structural,rlz_id
 0,8.51288E+02,0
 1,3.28829E+02,0
@@ -7,5 +7,6 @@ event_id,structural,rlz_id
 4,3.42458E+02,0
 5,2.52500E+02,0
 6,4.81053E+02,0
+7,0.00000E+00,0
 8,5.79551E+02,0
 9,5.03661E+02,0