… openquake/qa_tests_data/event_based_risk/case_4/expected/losses_by_event.csv