By using an AccumDict instead of a sparse matrix. Here are the figures for the Slovenia calculation (the speedup is 6x). 
```
# before
calc_100, maxmem=3.5 GB      time_sec memory_mb counts
============================ ======== ========= ======
total event_based_risk       977      163       40    
aggregating losses           629      0.0       5_240 
averaging losses             224      0.0       5_240 
computing risk               81       0.0       2_620 

# now
calc_98, maxmem=3.8 GB       time_sec memory_mb counts
============================ ======== ========= ======
total event_based_risk       540      182       40    
averaging losses             295      0.0       5_240 
aggregating losses           99       0.0       5_240 
computing risk               97       0.0       2_620 
```
For a more realistic calculation, like the one for Europe by Jamal, the speedup is a lot more impressive: 17x!